requirements.txt
streamlit
pandas
openmeteo-requests
requests-cache
retry-requests

